+++ GuassDePyramid.h
@@ -4,8 +4,8 @@
 
 #ifndef SIFT_GUASS_NORMAL_GUASSDEPYRAMID_H
 #define SIFT_GUASS_NORMAL_GUASSDEPYRAMID_H
-const double sigma = 2.0;
-const double PI = 3.1414926;
+const float sigma = 2.0;
+const float PI = 3.1414926;
 #include <iostream>
 #include <math.h>
 class GaussPyramid{
@@ -13,7 +13,7 @@ public:
     int** data; //记录图像灰度数据
     GaussPyramid();
     GaussPyramid(int** img, int len/*,int wid*/, int S);
-    double**** GaussPy;
+    float**** GaussPy;
     void GaussPyInit();
     void output();
     void GaussFilter(int theLayer);
@@ -24,7 +24,7 @@ protected:
     //int width;  先尝试宽度相同的版本
     int S;  //提取图像特征后,需要进行对比的图片数
     int layer;
-    double* filter;
+    float* filter;
 };
 
 GaussPyramid::GaussPyramid() {
@@ -49,19 +49,19 @@ GaussPyramid::GaussPyramid(int **img, int len, int S) {
         len/=2;
     }
     layer=x;
-    GaussPy=new double***[layer];
-    filter=new double[length];
+    GaussPy=new float***[layer];
+    filter=new float[length];
     GaussPyInit();
 }
 //初始化高斯金字塔，尚未进行高斯滤波操作
 void GaussPyramid::GaussPyInit() {
     int step=1;
     for (int i = 0; i < layer; ++i) {
-        GaussPy[i]=new double**[S+3];
+        GaussPy[i]=new float**[S+3];
         for (int j = 0; j < S + 3; ++j) {
-            GaussPy[i][j]=new double*[length/step];
+            GaussPy[i][j]=new float*[length/step];
             for (int k = 0; k < length/step; ++k) {
-                GaussPy[i][j][k]=new double [length/step];
+                GaussPy[i][j][k]=new float [length/step];
             }
         }
         step*=2;
@@ -99,7 +99,7 @@ void GaussPyramid::output() {
 }
 
 void GaussPyramid::GaussFilter(int theLayer) {//采用双边滤波
-    double len=length;
+    float len=length;
     int t=theLayer;
     while (theLayer!=0){
         theLayer--;
@@ -110,7 +110,7 @@ void GaussPyramid::GaussFilter(int theLayer) {//采用双边滤波
     len=(len-1)/2;
 
     for (int i = 0; i < S + 3; ++i) {
-        double sig=sigma/(i+1);
+        float sig=sigma/(i+1);
         for (int i = 0; i < MyLen; ++i) {
             filter[i] = exp(-(i-len)*(i-len)/(2*sig*sig))/(sigma*sqrt(2*PI));
         }
